fix bags applications

+++ app/api/applications/crud/application_crud.py
@@ -1,9 +1,12 @@
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
-from model.models import (Application, User, Advertisement, 
-                          UserRole, UserLogo, Location, Street,
-                          AdvertisementPhoto)
+from model.models import (
+    Application, User, Advertisement, 
+    UserRole, UserLogo, Location, Street,
+    AdvertisementPhoto
+)
+from fastapi import HTTPException
 import logging
 from typing import List
 
@@ -43,7 +46,17 @@ async def dal_get_applications_by_user_ads(user_id: int, db: AsyncSession) -> Li
             .selectinload(Advertisement.type_advertisement),
         selectinload(Application.advertisement)
             .selectinload(Advertisement.advertisement_photos)
-            .selectinload(AdvertisementPhoto.photo)
+            .selectinload(AdvertisementPhoto.photo),
+        selectinload(Application.advertisement)
+            .selectinload(Advertisement.favorites),
+        selectinload(Application.advertisement)
+            .selectinload(Advertisement.user_role)
+            .selectinload(UserRole.user)
+            .selectinload(User.user_logos)
+            .selectinload(UserLogo.logo),
+        selectinload(Application.advertisement)
+            .selectinload(Advertisement.user_role)
+            .selectinload(UserRole.role)
     )
     result = await db.execute(stmt)
     applications = result.scalars().all()
@@ -63,6 +76,41 @@ async def dal_delete_application_by_id(application_id: int, db: AsyncSession) ->
     stmt = select(Application).where(Application.id == application_id)
     result = await db.execute(stmt)
     application = result.scalar_one_or_none()
+    if not application:
+        logger.error(f"Application ID {application_id} not found for deletion")
+        raise HTTPException(status_code=404, detail="Отклик не найден")
+    
     await db.delete(application)
     await db.commit()
     logger.info(f"Application ID {application_id} deleted successfully")
+
+
+async def dal_get_applications_by_user(user_id: int, db: AsyncSession) -> List[Application]:
+    stmt = select(Application).where(Application.user_id == user_id).options(
+        selectinload(Application.user)
+            .selectinload(User.user_logos)
+            .selectinload(UserLogo.logo),
+        selectinload(Application.advertisement)
+            .selectinload(Advertisement.location)
+            .selectinload(Location.street)
+            .selectinload(Street.city),
+        selectinload(Application.advertisement)
+            .selectinload(Advertisement.type_advertisement),
+        selectinload(Application.advertisement)
+            .selectinload(Advertisement.advertisement_photos)
+            .selectinload(AdvertisementPhoto.photo),
+        selectinload(Application.advertisement)
+            .selectinload(Advertisement.favorites),
+        selectinload(Application.advertisement)
+            .selectinload(Advertisement.user_role)
+            .selectinload(UserRole.user)
+            .selectinload(User.user_logos)
+            .selectinload(UserLogo.logo),
+        selectinload(Application.advertisement)
+            .selectinload(Advertisement.user_role)
+            .selectinload(UserRole.role)
+    )
+    result = await db.execute(stmt)
+    applications = result.scalars().all()
+    logger.info(f"Retrieved {len(applications)} applications for user_id {user_id}")
+    return applications
